Move remaining rules out of typing category
