+++ fortitude/resources/test/fixtures/correctness/C051_ok.f90
