+++ fortitude/resources/test/fixtures/correctness/C052_ok.f90
